index/requirements.txt
flask>=3.1.0
requests>=2.32.3
python-dotenv>=1.1.0
redis>=6.2.0
celery>=5.5.2
gitpython>=3.1.44
llama-index>=0.12.37
llama-index-llms-ollama>=0.5.6
llama-index-vector-stores-qdrant>=0.6.0
qdrant-client>=1.14.2
llama-index-embeddings-huggingface>=0.5.4
